# Remove VersionEdit's friends pattern