@jowlyzhang 2023-10-30
> LGTM :) I'm not worried about perf regression where simple inlining takes care of optimizing

Thank you for the review. There is a crash test failure. https://app.circleci.com/pipelines/github/facebook/rocksdb/34096/workflows/31113b83-8b14-4a4d-a9d2-4ab5b4307577/jobs/696392 I think it's not related to this change. I don't have proof yet, but by checking recent commit history. I think this change https://github.com/facebook/rocksdb/blob/76402c034e8fa75809df3ed89a5742f044e980e9/db_stress_tool/db_stress_test_base.cc#L1931 #12015  could be related, because it's setting the Options* to an object on the stack that is gonna be destroyed upon the function returns.